requirements.txt
streamlit
numpy
pillow
opencv-python
opencv-python-headless

SpeechRecognition
load-dotenv
langchain
langchain-groq
langchain-google-genai
langchain-core
langchain-community

fpdf
tensorflow
tf-keras
language-tool-python
pdf2image
scikit-learn
pypdf
faiss-cpu
